requirements.txt
langchain-anthropic==1.3.5
langchain-core==1.2.19
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyYAML==6.0.3
tenacity==9.1.4
python-dotenv==1.2.2